+++ packages/contracts/solana-spoke/programs/everclear_spoke/src/error.rs
@@ -64,4 +64,10 @@ pub enum SpokeError {
     InvalidSeeds,
     #[msg("Invalid vault account")]
     InvalidVaultAccount,
+    #[msg("Invalid intent pda")]
+    InvalidIntentPda,
+    #[msg("Invalid settlement size")]
+    InvalidSettlementSize,
+    #[msg("Incorrect settlement accounts, mismatch with intent PDA.")]
+    IncorrectSettlementAccounts,
 }
