+++ packages/contracts/solana-spoke/programs/everclear_spoke/src/state.rs
@@ -1,6 +1,9 @@
 use anchor_lang::prelude::*;
 
-use crate::hyperlane::InterchainGasPaymasterType;
+use crate::{
+    hyperlane::{InterchainGasPaymasterType, SerializableAccountMeta},
+    instructions::messages::Settlement,
+};
 
 /// SpokeState – global configuration.
 #[account]
@@ -23,8 +26,6 @@ pub struct SpokeState {
     pub nonce: u64,
     // Owner of the program (admin).
     pub owner: Pubkey,
-    // Intent status mapping.
-    pub status: Vec<IntentStatusAccount>,
     // Bump for PDA.
     pub bump: u8,
     // Mailbox address
@@ -48,7 +49,6 @@ impl SpokeState {
         + 8                      // message_gas_limit: u64
         + 8                      // nonce: u64
         + 32                     // owner: Pubkey
-        + 4                      // status HashMap
         + 1                      // bump: u8
         + 32                     // mailbox: Pubkey
         + 1                      // mailbox_dispatch_authority_bump: u8
@@ -62,6 +62,8 @@ impl SpokeState {
 pub struct IntentStatusAccount {
     pub key: [u8; 32],
     pub status: IntentStatus,
+    pub settlement: Option<Settlement>,
+    pub accounts: Vec<SerializableAccountMeta>,
 }
 
 /// Intent status.
@@ -72,4 +74,5 @@ pub enum IntentStatus {
     Filled,
     Settled,
     SettledAndManuallyExecuted,
+    Delivered,
 }
